refactor(data, game-logic): isolate input handling and game loop
Separate receiving and parsing into helper functions and move them to DataUtils file
Isolate the main game loop into a game function and place it in GameLogic file to improve separation of concerns
