+++ src/main/kotlin/racingcar/Application.kt
@@ -1,33 +1,8 @@
 package racingcar
-import camp.nextstep.edu.missionutils.Console
-
-fun setCarData(input: String): MutableMap<String, String> {
-    val carsList = input.split(",")
-    val carsMap: MutableMap<String, String> = mutableMapOf()
-
-    for (car in carsList) {
-        carsMap[car] = ""
-    }
-    return carsMap
-}
 
 fun main() {
-    println("Enter the names of the cars (comma-separated):")
-    val carNamesInput: String = Console.readLine()
-    validateCarsInput(input = carNamesInput)
+    val cars = getCars()
+    val rounds = getNumberOfRounds()
 
-    println("How many rounds will be played?")
-    val roundsInput = Console.readLine()
-    validateRoundsInput(input = roundsInput)
-
-    var rounds: Int = roundsInput.toInt()
-    val cars: MutableMap<String, String> = setCarData(carNamesInput)
-
-    println("Race Results")
-    while (rounds > 0) {
-        runRound(cars)
-        println()
-        rounds--
-    }
-    printWinners(cars)
+    game(cars, rounds);
 }
